Use `/` as the default status endpoint
It seems strange to have the default status endpoint set as
`/status/check`, as most people do not implement this as an endpoint in
their web applications. The `/status/check` endpoint is more of an
internal convention than a community standard, so we should default this
to something more likely to be able to report an application's health.
In most cases, the root path should simply work for this out of the box.

Signed-off-by: David Celis <me@davidcel.is>