Add 'command' DSL method for customizable Cmd param
This fixes #19. It's a bit of a messy change, since I took a couple
stylistic liberties while implementing the feature.

Primarily, the intent for this feature is to support customizable
commands, because not every Docker image is built with the metadata
describing how it should be started. And sometimes users may wish to
start the same image in multiple ways.

This adds a DSL method `command`, as such:

  task :staging do
    host 'docker-server-1.example.com'
    command ['/bin/bash', '-c', '/bin/echo hi']
  end

Along the way, this refactors a couple of the lib/centurion/deploy.rb
methods to use default parameters that are closer to the data they
describe. And this also adjusts some RSpec tests to use `let` blocks so
that the tests are more readable (I found it very difficult to figure
out which parameters all the {}'s and []'s were referring to.)