Handle Docker registry API inconsistency
I just opened an issue here[1], but since this is a bug biting people
now, I figure it's worth a fix.

This is caused because if the user is using the official Docker registry
to host their images, the API response for /v1/repositories/:repo/tags
is quite different than the current open-source registry[2]'s response.

This commit / PR adds a standardization based on the URL of the
registry. It's not perfect, but it seems likely to work for most folks.

[1]: https://github.com/docker/docker-registry/issues/517
[2]: https://github.com/docker/docker-registry