Offer dry-run capability for centurion
Offer a dry-run possibility. This would allow people who want to try
customizing a docker command on a host to acquire all environment
variables and other configuration and then run docker from command line.

It's useful if they want to run a different command that is not in their
rake file, mount a volume not defined yet, or even just tweak a bit what
they have.