Adds a method for finding the correct server to canary to.
This will be used in an upcoming PR to do canary deployments.  There are two scenarios where this will work.  First, if all of the servers have the same docker image deployed to them then it will just pick the first server as the canary.  If all of the servers have the same docker image deployed to them except for one which has a different image then it will choose that one as the canary, assuming that it had a previous canary already deployed to it.

The next step is to add a method that gets the current non-canary deployment image.  Then we can do nifty things like automatically rolling back a canary to align with the rest of the environment or promoting one environment's image (staging) to another (production).

Also fixes a deprecation warning in another test.