don't stop all containers on deploy_console
deploy_console is suposed to deploy a console on the first host
declared on the centurion config, but it was actually doing it
on every host.

The new behavior is to deploy the console only on the first host
instead of stoping the running container for running a new one
with the console it will launch a new container which will run
without interrumpting the service, to do that it clears the
PortsBindings in the console container to prevent port conflicts.