fixed cgroup setting bug
Cgroup settings ceased to be set after the most recent refactor (and introduction of the Service class).
Was just a matter of the values not being fetched and set (in Service::self.from_env and Service::build_host_config)

Also added 2 new tests for mem/cpu share limits and removed a broken dogestry test.