Expand healthchecks to include a string type
health_check is currently restricted to being a callable. This
removes that restriction and also adds an health_check_args to
the DSL.

We will want to pass the values of health_check and
health_check_args to Marathon as labels, so the executor can
know how to health check its container.

Backwards compatability with the existing scheme of passing a
callable for use by Centurion itself is mainted.