Pass stop_timeout down to underlying http request
When setting a custom stop_timeout, the value is passed to the docker API, but Excon won't wait any longer than 120 seconds for the response. This can lead to incorrect timeout errors if the container takes longer than that to stop.

This passes the docker stop_timeout to the Excon request so it will wait at least as long as docker will for the container to stop.